they dont want reservations. what do they want?
they dont know what they want. so they get you to figure that out for
them. nopt only do you have to do their shit, you also have to read
their thoughts to understand how they want this shit? gotta get used to
this. its gna be my reality of things for a while. "if one has the
mental capacity to understand and help, one is responsible to do so"...